refactor: 빈 등록 역할 BeanRegister 로 위임
