+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handleradaptor/HandlerAdaptors.java
@@ -1,7 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.tobe.handleradaptor;
 
-import com.interface21.container.BeanContainer;
-import com.interface21.scanner.BeanScanner;
+import com.interface21.bean.container.BeanContainer;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdaptor;
 import java.util.List;
 
@@ -10,10 +9,7 @@ public class HandlerAdaptors {
     private final List<HandlerAdaptor> handlerAdaptors;
 
     public HandlerAdaptors() {
-        String packageName = getClass().getPackageName();
-        List<Object> beans = BeanScanner.subTypeScan(HandlerAdaptor.class, packageName);
         BeanContainer beanContainer = BeanContainer.getInstance();
-        beanContainer.registerBean(beans);
         this.handlerAdaptors = beanContainer.getSubTypeBeansOf(HandlerAdaptor.class);
     }
