+++ mvc/src/test/java/com/interface21/container/BeanContainerTest.java
@@ -3,6 +3,7 @@ package com.interface21.container;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import com.interface21.bean.container.BeanContainer;
 import com.interface21.context.stereotype.Controller;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,7 +29,7 @@ class BeanContainerTest {
     @DisplayName("특정 애너테이션이 포함된 빈을 찾을 수 있다.")
     @Test
     void getAnnotatedBeans() {
-        beanContainer.registerBean(List.of(new TestBean1(), new TestBean2(), new TestBean3()));
+        beanContainer.registerBeans(List.of(new TestBean1(), new TestBean2(), new TestBean3()));
         List<Object> annotatedBeans = beanContainer.getAnnotatedBeans(Controller.class);
 
         assertThat(annotatedBeans).hasSize(2)
@@ -39,7 +40,7 @@ class BeanContainerTest {
     @DisplayName("특정 클래스를 구현한 빈을 찾을 수 있다.")
     @Test
     void getSubTypeBeansOf() {
-        beanContainer.registerBean(List.of(new TestBean1(), new TestBean2(), new TestBean3()));
+        beanContainer.registerBeans(List.of(new TestBean1(), new TestBean2(), new TestBean3()));
         List<TestBean2> annotatedBeans = beanContainer.getSubTypeBeansOf(TestBean2.class);
 
         assertThat(annotatedBeans).hasSize(2)
@@ -50,7 +51,7 @@ class BeanContainerTest {
     @DisplayName("컨테이너는 객체를 싱글톤으로 관리한다.")
     @Test
     void singleTon() {
-        beanContainer.registerBean(List.of(new BeanContainerTest()));
+        beanContainer.registerBeans(List.of(new BeanContainerTest()));
         Object bean1 = beanContainer.getBean(BeanContainerTest.class);
         Object bean2 = beanContainer.getBean(BeanContainerTest.class);
