+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/tobe/handlermapping/HandlerMappingsTest.java
@@ -2,7 +2,7 @@ package com.interface21.webmvc.servlet.mvc.tobe.handlermapping;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.interface21.container.BeanContainer;
+import com.interface21.bean.container.BeanContainer;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
 import jakarta.servlet.http.HttpServletRequest;
